[FIX] UIManager
UI 매니저 기능 분할로 새 클래스 생성 및 게임매니저 수정
