+++ Assets/Worker/NGH/Scripts/GameManager.cs
@@ -117,53 +117,12 @@ public class GameManager : MonoBehaviour
     // 씬 전환후 연결 초기화
     private void InitializeBattleScene()
     {
-        UIManager.Instance.lobbyScreen = null;
-        UIManager.Instance.magicShopScreen = null;
-        UIManager.Instance.statusUpgradeShopScreen = null;
+        
     }
 
     private void InitializeLobbyScene()
     {
-        // 인스펙터 상 각 오브젝트 연결
-        UIManager.Instance.lobbyScreen = GameObject.FindWithTag("Lobby");
-        GameObject[] allObjects = Resources.FindObjectsOfTypeAll<GameObject>();
-
-        foreach (GameObject obj in allObjects)
-        {
-            if (obj.CompareTag("Magicshop"))
-            {
-                UIManager.Instance.magicShopScreen = obj;
-            }
-            if (obj.CompareTag("Statusshop"))
-            {
-                UIManager.Instance.statusUpgradeShopScreen = obj;
-            }
-        }
-        Button[] allButtons = Resources.FindObjectsOfTypeAll<Button>();
-        foreach (Button button in allButtons)
-        {
-            if(button.CompareTag("Gamestart"))
-            {
-                UIManager.Instance.gameStartButton = button;
-            }
-            if (button.CompareTag("Magicshopenter"))
-            {
-                UIManager.Instance.magicShopButton = button;
-            }
-            if(button.CompareTag("Statusshopenter"))
-            {
-                UIManager.Instance.statusUpgradeShopButton = button;
-            }
-            if(button.CompareTag("Magicshopexit"))
-            {
-                UIManager.Instance.magicShopExit = button;
-            }
-            if (button.CompareTag("Statusshopexit"))
-            {
-                UIManager.Instance.statusUpgradeShopExit = button;
-            }
-        }
-        UIManager.Instance.LinkButton();
+        
     }
 
     // 골드를 추가하는 메서드
