+++ Assets/Worker/NGH/Scripts/StatusUpgradeShopUI.cs
@@ -2,7 +2,7 @@ using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
 
-public class MagicShop : MonoBehaviour
+public class StatusUpgradeShopUI : MonoBehaviour
 {
     // Start is called before the first frame update
     void Start()
