+++ Assets/Worker/NGH/Scripts/StatusUpgradeShopUI.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 6ec6146f0e58124439c7b170d86c3983
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
